Don't cache input shapes because they can vary
Input shape transforms were previously only calculated once (i.e. for
one frame / one image) and then cached for all subsequent calls.
This is fine in 99% of use cases but if you want to process differently
shaped inputs in the same deface process (given multiple input paths),
this leads to crashes or even silently incorrect outputs.

Transforms now have to be recalculated for each frame but this is
very cheap, so it's okay to do this for the sake of better stability.

Fixes #41